refactor: externalize component PreventLeavingPageByMistake
